platform/machine/qemu: centralize qemu code
Prepares to add the unprivileged-qemu platform by moving the qemu code
to a centralized location.
